Detect mismatches between throw spec and thrown exceptions.
That is, we detect:
- if a function throws an exception, but does not document it
- if a function documents an exception, but does not throw it
- if an invalid exception is thrown (non-object)

There are a couple of things to consider, though:
- if a function calls a function that throws, but does not
document it, PHPCheck does not complain
- if a non-abstract method is overridden and the PHPDOc for
the parent method says that it throws because some of his
childs might, PHPCheck complains. This might be a valid
way of documenting exceptions, but I think in most
cases this will be an error. The problem is that there
can always be subclasses (not existing at compile time)
that override the method to make it okay in this sense.
So, in theory, we are never absolutely sure whether it's
okay or not.