Store variables and display them properly.
Although it might be a bit confusing that the type/value
of the variable is always the one at the end of a function
scope. Therefore, if a variable is defined in, e.g., just
one branch of an if-else statement, it will be undefined
afterwards. But I think, displaying them is still helpful.