Detected unused variables and parameters.
Parameters are handled special, because it happens often
that a subclass overrides a method and does not use all
parameters. But it has to specify them nevertheless.
Therefore, we don't report an unused parameter in overridden
methods. However, we still report it for methods that have
parameters that they don't need themselve, but subclasses
that override the method might. But this seems acceptable.