Major refactoring.
I moved all the analyzing stuff out of the statement scanner
into corresponding classes.

Furhermore, I introduced an environment class that holds all
stuff that we need, i.e., types, errors, options and the type
storage. This is passed around to reduce the number of
parameters.

And finally, I moved the analyzer step into the statement
scanner, so that we have just two phases now. The last one
wasn't really necessary both technically and for the user.
Because it didn't take enough of time to justify running
it separately. The report-{unknown,mixed} settings have
been moved into the project settings.