Removed report-unknown and changed report-mixed.
I noticed that now that the bug with the classname
handling is fixed, we have much more false positives,
because one possible type does not conform to the spec.
This is because it happens quite often that e.g., a
function returns a string or false and the caller
exits early in case of false. But PHPCheck does still
complain, because it does not see that the code exits
early. Thus, I decided to introduce an option that
controls whether these types of potential errors
should be reported or not.

Furthermore, I removed report-unknown and replaced
report-mixed with this new option. Report-unknown
does actually never make a lot of sense, because it
reports basically only false positives.