Support references.
That is, we now detect reference parameters in the PHP manual
and analyzed code and implicitly create a variable for calls where
the variable does not exist already.
Unfortunately, it is a bit difficult to implement properly,
because when calling a function we only now the type, but now
where it came from (e.g., a variable). But a small hack makes
that possible now :)

Furthermore, we now don't report an error if a reference parameter
is only written. And we report errors if the PHPDoc of a parameter
or return value does not match the parameter/return. That is, if
type hinting and '&' in front of a parameter differs from the
PHPDoc, we complain.