Let projects depend on other projects.
If using a library or framework, one can now add a separate
project for the library/framework and let the actual project
depend on this. This way, one can work on the actual project
and reparse it often without needing to reparse the library/
framework over and over again as well.

Basically, we already had a limited version of this: all
projects implicitly depended on the PHP builtins (and still
do).